Upgrade cardano-common-explorer to version 1.0.6

+++ src/main/java/com/cardano/explorer/model/response/TxFilterResponse.java
@@ -1,7 +1,7 @@
 package com.cardano.explorer.model.response;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.time.LocalDateTime;
 import java.util.List;
 import lombok.Getter;
@@ -20,7 +20,7 @@ public class TxFilterResponse implements Serializable {
   private LocalDateTime time;
   private List<String> addressesInput;
   private List<String> addressesOutput;
-  private BigDecimal fee;
-  private BigDecimal totalOutput;
+  private BigInteger fee;
+  private BigInteger totalOutput;
 
 }
