+++ src/main/java/com/cardano/explorer/model/response/address/AddressResponse.java
@@ -1,7 +1,7 @@
 package com.cardano.explorer.model.response.address;
 
 import com.cardano.explorer.model.response.token.TokenAddressResponse;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -17,7 +17,7 @@ import lombok.Setter;
 public class AddressResponse {
   private String address;
   private Long txCount;
-  private BigDecimal balance;
+  private BigInteger balance;
   private List<TokenAddressResponse> tokens;
   private String stakeAddress;
   private Boolean isContract;
