+++ src/main/java/com/cardano/explorer/repository/EpochStakeRepository.java
@@ -4,7 +4,7 @@ import com.cardano.explorer.model.response.pool.projection.EpochChartProjection;
 import com.cardano.explorer.model.response.pool.projection.EpochStakeProjection;
 import com.cardano.explorer.projection.StakeAddressProjection;
 import com.sotatek.cardano.common.entity.EpochStake;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 import java.util.Optional;
 import java.util.Set;
@@ -23,7 +23,7 @@ public interface EpochStakeRepository extends JpaRepository<EpochStake, Long> {
       @Param("poolId") Long poolId);
 
   @Query(value = "SELECT sum(es.amount) FROM EpochStake es WHERE es.epochNo = :epochNo")
-  Optional<BigDecimal> totalStakeAllPoolByEpochNo(@Param("epochNo") Integer epochNo);
+  Optional<BigInteger> totalStakeAllPoolByEpochNo(@Param("epochNo") Integer epochNo);
 
   @Query(value =
       "SELECT es.epochNo AS chartKey, sum(es.amount) AS chartValue FROM EpochStake es "
